# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Traffic Assistant Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /crowdsourcing/reports:
    get:
      operationId: listCrowdsourcingReports
      tags: [crowdsourcing]
      summary: List crowdsourcing reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrowdsourcingReportList"
  /crowdsourcing/reports/{id}:
    get:
      operationId: getCrowdsourcingReport
      tags: [crowdsourcing]
      summary: Get a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrowdsourcingReport"
    patch:
      operationId: updateCrowdsourcingReport
      tags: [crowdsourcing]
      summary: Approve or reject a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCrowdsourcingReportBody"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrowdsourcingReport"
  /crowdsourcing/sign-reports:
    get:
      operationId: listSignReports
      tags: [crowdsourcing]
      summary: List sign error/missing reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, hidden, visible]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of sign reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignReportList"
  /crowdsourcing/sign-reports/{id}:
    patch:
      operationId: updateSignReport
      tags: [crowdsourcing]
      summary: Toggle sign visibility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSignReportBody"
      responses:
        "200":
          description: Updated sign report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignReport"
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List rule engine rules
      parameters:
        - name: vehicleType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleList"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleBody"
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /rules/{id}:
    get:
      operationId: getRule
      tags: [rules]
      summary: Get a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rule detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete a rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List admins
      responses:
        "200":
          description: List of admins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminList"
    post:
      operationId: createAdmin
      tags: [admins]
      summary: Create an admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminBody"
      responses:
        "201":
          description: Created admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
  /admins/{id}:
    patch:
      operationId: updateAdmin
      tags: [admins]
      summary: Update admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminBody"
      responses:
        "200":
          description: Updated admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
    delete:
      operationId: deleteAdmin
      tags: [admins]
      summary: Delete admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List AI models
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
    post:
      operationId: createModel
      tags: [models]
      summary: Upload a new model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelBody"
      responses:
        "201":
          description: Created model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
  /models/{id}/deploy:
    post:
      operationId: deployModel
      tags: [models]
      summary: Deploy model OTA to all devices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deploy started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
...
components:
  schemas:
...
    CrowdsourcingReport:
      type: object
      properties:
        id:
          type: integer
        lat:
          type: number
        lng:
          type: number
        imageUrl:
          type: ["string", "null"]
        signCode:
          type: string
        signType:
          type: string
        confidenceScore:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        submittedBy:
          type: string
        submittedAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - lat
        - lng
        - signCode
        - signType
        - confidenceScore
        - status
        - submittedBy
        - submittedAt
    CrowdsourcingReportList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CrowdsourcingReport"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
    UpdateCrowdsourcingReportBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        notes:
          type: string
      required:
        - status
    SignReport:
      type: object
      properties:
        id:
          type: integer
        signId:
          type: integer
        signCode:
          type: string
        lat:
          type: number
        lng:
          type: number
        reportCount:
          type: integer
        reportType:
          type: string
          enum: [missing, error]
        visibility:
          type: string
          enum: [visible, hidden]
        lastReportedAt:
          type: string
      required:
        - id
        - signId
        - signCode
        - lat
        - lng
        - reportCount
        - reportType
        - visibility
        - lastReportedAt
    SignReportList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SignReport"
        total:
          type: integer
      required:
        - items
        - total
    UpdateSignReportBody:
      type: object
      properties:
        visibility:
          type: string
          enum: [visible, hidden]
      required:
        - visibility
...
    Rule:
      type: object
      properties:
        id:
          type: integer
        signCode:
          type: string
        signContent:
          type: string
        vehicleType:
          type: string
          enum: [car, truck, motorcycle, bicycle, all]
        warningMessage:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - signCode
        - signContent
        - vehicleType
        - warningMessage
        - isActive
        - createdAt
        - updatedAt
    RuleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Rule"
        total:
          type: integer
      required:
        - items
        - total
    CreateRuleBody:
      type: object
      properties:
        signCode:
          type: string
        signContent:
          type: string
        vehicleType:
          type: string
          enum: [car, truck, motorcycle, bicycle, all]
        warningMessage:
          type: string
        isActive:
          type: boolean
      required:
        - signCode
        - signContent
        - vehicleType
        - warningMessage
    UpdateRuleBody:
      type: object
      properties:
        signCode:
          type: string
        signContent:
          type: string
        vehicleType:
          type: string
          enum: [car, truck, motorcycle, bicycle, all]
        warningMessage:
          type: string
        isActive:
          type: boolean
    Admin:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        district:
          type: string
        geoPolygon:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, district_admin]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - district
        - role
        - isActive
        - createdAt
    AdminList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Admin"
        total:
          type: integer
      required:
        - items
        - total
    CreateAdminBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        district:
          type: string
        geoPolygon:
          type: string
        role:
          type: string
          enum: [super_admin, district_admin]
      required:
        - name
        - email
        - district
        - role
    UpdateAdminBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        district:
          type: string
        geoPolygon:
          type: string
        role:
          type: string
          enum: [super_admin, district_admin]
        isActive:
          type: boolean
...
    Model:
      type: object
      properties:
        id:
          type: integer
        version:
          type: string
        filename:
          type: string
        releaseNote:
          type: string
        status:
          type: string
          enum: [draft, deploying, deployed, failed]
        deployedAt:
          type: ["string", "null"]
        deviceCount:
          type: integer
        uploadedAt:
          type: string
      required:
        - id
        - version
        - filename
        - releaseNote
        - status
        - deviceCount
        - uploadedAt
    ModelList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Model"
        total:
          type: integer
      required:
        - items
        - total
    CreateModelBody:
      type: object
      properties:
        version:
          type: string
        filename:
          type: string
        releaseNote:
          type: string
      required:
        - version
        - filename
        - releaseNote
...